% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP4LCF-package.R
\docType{package}
\name{DGP4LCF-package}
\alias{DGP4LCF}
\alias{DGP4LCF-package}
\title{DGP4LCF: Modeling Interactive Biological Pathways From High-Dimensional Gene Expression Trajectories Using Dependent Gaussian Processes}
\description{
This package implements the BSFA-DGP model proposed in the paper.
}
\keyword{internal}
