% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_loading_heatmap.R
\name{factor_loading_heatmap}
\alias{factor_loading_heatmap}
\title{Displaying significant factor loadings in the heatmap.}
\usage{
factor_loading_heatmap(factor_loading_matrix, heatmap_title)
}
\arguments{
\item{factor_loading_matrix}{A matrix of dimension (p, k), which stores results for factor loadings.}

\item{heatmap_title}{A character. Title for the heatmap.}
}
\value{
A heatmap presenting posterior median estimates of factor loadings.
}
\description{
This function is used to visualize results of estimates of factor loadings (in heatmaps).
}
\examples{
# See examples in vignette
vignette("bsfadgp_regular_data_example", package = "DGP4LCF")

}
