% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_score_trajectory.R
\name{factor_score_trajectory}
\alias{factor_score_trajectory}
\title{Plotting figures for factor score trajectory.}
\usage{
factor_score_trajectory(
  factor_score_matrix,
  factor_index,
  person_index,
  trajectory_title,
  cex_main = 1
)
}
\arguments{
\item{factor_score_matrix}{A matrix of dimension (q, k, n), used to store results for factor scores.}

\item{factor_index}{A numeric scalar. Index of the factor of interest.}

\item{person_index}{A numeric scalar. Index of the person of interest.}

\item{trajectory_title}{A character. Title for the factor trajectory plot.}

\item{cex_main}{A numeric scalar. Text size of the title.}
}
\value{
Trajectory of the designated person-factor.
}
\description{
This function is used to visualize results of factor score trajectories.
}
\examples{
# See examples in vignette
vignette("bsfadgp_regular_data_example", package = "DGP4LCF")

}
