% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_after_mcem_combine_chains.R
\name{gibbs_after_mcem_combine_chains}
\alias{gibbs_after_mcem_combine_chains}
\title{Combining from all chains the posterior samples for parameters in the model and predicted gene expressions.}
\usage{
gibbs_after_mcem_combine_chains(tot_chain, gibbs_after_mcem_algorithm_result)
}
\arguments{
\item{tot_chain}{A numeric scalar. Total number of chains.}

\item{gibbs_after_mcem_algorithm_result}{A list of objects storing model constants. Should be the same as that input to the 'function gibbs_after_mcem_load_chains'.}
}
\value{
All saved posterior samples for parameters in the model and predicted gene expressions.
}
\description{
Combining from all chains the posterior samples for parameters in the model and predicted gene expressions.
}
\examples{
# See examples in vignette
vignette("bsfadgp_regular_data_example", package = "DGP4LCF")

}
