% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_after_mcem_diff_initials.R
\name{gibbs_after_mcem_diff_initials}
\alias{gibbs_after_mcem_diff_initials}
\title{Generating different initials for multiple chains.}
\usage{
gibbs_after_mcem_diff_initials(
  ind_x = TRUE,
  tot_chain = 5,
  mcem_parameter_setup_result,
  mcem_algorithm_result
)
}
\arguments{
\item{ind_x}{A logical value. ind_x = TRUE uses the model including the intercept term for subject-gene mean in after-MCEM-Gibbs sampler; otherwise uses the model without the intercept term.}

\item{tot_chain}{A numeric scalar. Number of parallel chains.}

\item{mcem_parameter_setup_result}{A list of objects returned from the function 'mcem_parameter_setup'.}

\item{mcem_algorithm_result}{A list of objects returned from the function 'mcem_algorithm'.}
}
\value{
Different initials for multiple chains.
}
\description{
Generating different initials for multiple chains.
}
\examples{
# See examples in vignette
vignette("bsfadgp_regular_data_example", package = "DGP4LCF")
vignette("bsfadgp_irregular_data_example", package = "DGP4LCF")

}
