% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_after_mcem_load_chains.R
\name{gibbs_after_mcem_load_chains}
\alias{gibbs_after_mcem_load_chains}
\title{Loading the saved posterior samples for parameters in the model and predicted gene expressions.}
\usage{
gibbs_after_mcem_load_chains(chain_index, gibbs_after_mcem_algorithm_result)
}
\arguments{
\item{chain_index}{A numeric scalar. Index of the chain.}

\item{gibbs_after_mcem_algorithm_result}{A list of objects storing model constants.}
}
\value{
All saved posterior samples for parameters in the model and predicted gene expressions.
}
\description{
Loading the saved posterior samples for parameters in the model and predicted gene expressions.
}
\examples{
# See examples in vignette
vignette("bsfadgp_regular_data_example",  package = "DGP4LCF")

}
