% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcem_cov_plot.R
\name{mcem_cov_plot}
\alias{mcem_cov_plot}
\title{Visualizing cross-correlations among factors.}
\usage{
mcem_cov_plot(k, q, cov_input, title)
}
\arguments{
\item{k}{A numeric scalar. Number of latent factors.}

\item{q}{A numeric scalar. Number of time points in the covariance matrix of factors.}

\item{cov_input}{A matrix of dimension (kq, kq). The covariance matrix of the vector obtained from vectorizing the matrix of latent factor scores.}

\item{title}{A character. Title for the plot.}
}
\value{
Visualization of cross-correlations among factors.
}
\description{
Visualizing cross-correlations among factors.
}
\examples{
# See examples in vignette
vignette("bsfadgp_regular_data_example", package = "DGP4LCF")
vignette("bsfadgp_irregular_data_example", package = "DGP4LCF")

}
