% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_fcs_init}
\alias{sim_fcs_init}
\title{Initials values.}
\format{
An object of class \code{list} of length 14.
}
\usage{
sim_fcs_init
}
\description{
Initial values provided by the two-step approach.
}
\keyword{datasets}
