% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_fcs_results_regular_8}
\alias{sim_fcs_results_regular_8}
\title{Results when people are observed at common 8 time points.}
\format{
An object of class \code{list} of length 3.
}
\usage{
sim_fcs_results_regular_8
}
\description{
Results when people are observed at common 8 time points.
}
\keyword{datasets}
