% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_fcs_truth}
\alias{sim_fcs_truth}
\title{Truth of simulated data.}
\format{
An object of class \code{list} of length 19.
}
\usage{
sim_fcs_truth
}
\description{
Simulated data under the scenario where factors are correlated and have small variability (CS).
}
\keyword{datasets}
