% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_generator.R
\name{table_generator}
\alias{table_generator}
\title{Generating a table listing all possible combinations of the binary variables for one gene.}
\usage{
table_generator(k)
}
\arguments{
\item{k}{A numeric scalar. Number of latent factors.}
}
\value{
A table listing all possible combinations of the binary variables for one gene.
}
\description{
Generating a table listing all possible combinations of the binary variables for one gene.
}
\examples{
# See examples in vignette
vignette("bsfadgp_regular_data_example", package = "DGP4LCF")
vignette("bsfadgp_irregular_data_example", package = "DGP4LCF")

}
