% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP4LCF-package.R
\docType{package}
\name{DGP4LCF-package}
\alias{DGP4LCF}
\alias{DGP4LCF-package}
\title{DGP4LCF: Dependent Gaussian Processes for Longitudinal Correlated Factors}
\description{
This package implements a novel methodology to model high-dimensional gene expression trajectories.
}
\author{
\strong{Maintainer}: 

Other contributors:
\itemize{
  \item Jiachen Cai \email{jiachen.cai@mrc-bsu.cam.ac.uk}
}

}
\keyword{internal}
